anonyfiles-api/requirements.txt
fastapi==0.111.0 # ou la version que vous préférez
uvicorn[standard]==0.29.0 # ou la version que vous préférez
python-multipart==0.0.9 # Version compatible pour UploadFile
pyyaml # pour load_config